refactor extract users accounts writeJSON from app.js to data.js
